requirements.txt
Django>=5.2,<5.3
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.4
django-axes>=7.0
django-auditlog>=3.0
django-environ>=0.11
